@@ -1,7 +1,7 @@
 package com.robertlevonyan.components.picker
 
-import android.support.annotation.LayoutRes
-import android.support.v7.widget.RecyclerView
+import androidx.annotation.LayoutRes
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 
@@ -9,7 +9,7 @@ internal class BaseAdapter<I>(private val items: List<I>,
                      @LayoutRes private val lyt: Int,
                      private val bind: View.(I, Int) -> Unit,
                      private val itemClick: (I, Int) -> Unit = { _, _ -> })
-    : RecyclerView.Adapter<BaseAdapter.BaseViewHolder>() {
+    : androidx.recyclerview.widget.RecyclerView.Adapter<BaseAdapter.BaseViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): BaseViewHolder {
         val v = parent inflate lyt
@@ -25,5 +25,5 @@ internal class BaseAdapter<I>(private val items: List<I>,
         holder.itemView.setOnClickListener { itemClick(item, position) }
     }
 
-    class BaseViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
+    class BaseViewHolder(itemView: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(itemView)
 }
